refactor : Controller API/Page 분리 및 주요 로직 Service 추가 및 이관
