+++ src/main/java/com/app/keywordwatcher/web/controller/auth/AuthPageController.java
@@ -1,0 +1,20 @@
+package com.app.keywordwatcher.web.controller.auth;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+
+@RequestMapping("/public")
+@Controller
+public class AuthPageController {
+    @GetMapping("/login")
+    public String loginPage() {
+        return "auth/login";
+    }
+
+    @GetMapping("/signup")
+    public String signupPage() {
+        return "auth/signup";
+    }
+}
