+++ src/main/java/com/app/keywordwatcher/web/controller/home/HomeController.java
@@ -1,0 +1,13 @@
+package com.app.keywordwatcher.web.controller.home;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping
+    public String home() {
+        return "index";
+    }
+}
