+++ src/main/java/com/app/keywordwatcher/web/controller/user/UserController.java
@@ -1,0 +1,16 @@
+package com.app.keywordwatcher.web.controller.user;
+
+import com.app.keywordwatcher.web.controller.ApiResponse;
+import com.app.keywordwatcher.web.dto.UserResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+
+    @GetMapping("/api/user")
+    public ApiResponse<UserResponse> getUserInfo(Authentication authentication) {
+        return ApiResponse.success(UserResponse.create(authentication.getName()));
+    }
+}
