+++ src/main/java/com/app/keywordwatcher/web/exception/LoginException.java
@@ -1,0 +1,15 @@
+package com.app.keywordwatcher.web.exception;
+
+public class LoginException extends RuntimeException {
+    public LoginException(String message) {
+        super(message);
+    }
+
+    public LoginException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public LoginException(Throwable cause) {
+        super(cause);
+    }
+}
