+++ src/main/java/com/app/keywordwatcher/web/service/AuthService.java
@@ -1,0 +1,72 @@
+package com.app.keywordwatcher.web.service;
+
+import com.app.keywordwatcher.domain.user.User;
+import com.app.keywordwatcher.domain.user.UserRepository;
+import com.app.keywordwatcher.web.dto.LoginRequest;
+import com.app.keywordwatcher.web.dto.SignupRequest;
+import com.app.keywordwatcher.web.exception.LoginException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class AuthService {
+    private final UserRepository userRepository;
+    private final AuthenticationManager authenticationManager;
+    private final PasswordEncoder passwordEncoder;
+
+    public Long signUp(SignupRequest request) {
+        if (request.getPassword() == null) {
+            throw new IllegalArgumentException("비밀번호를 입력해주세요.");
+        }
+
+        if (request.isPasswordNonMatch()) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
+        if (userRepository.existsByUserId(request.getUserId())) {
+            throw new IllegalArgumentException("이미 사용 중인 ID입니다.");
+        }
+
+        if (userRepository.existsByEmail(request.getEmail())) {
+            throw new IllegalArgumentException("이미 사용 중인 이메일입니다.");
+        }
+
+        return userRepository.save(User.create(request, passwordEncoder)).getId();
+    }
+
+    public void login(LoginRequest request, HttpServletRequest httpRequest) {
+        try {
+            Authentication authentication = authenticationManager.authenticate(
+                    new UsernamePasswordAuthenticationToken(request.getUsername(), request.getPassword())
+            );
+
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+
+            HttpSession session = httpRequest.getSession(true);
+            session.setAttribute("SPRING_SECURITY_CONTEXT", SecurityContextHolder.getContext());
+        } catch (RuntimeException e) {
+            log.error(e.getMessage());
+            throw new LoginException("로그인에 실패했습니다.");
+        }
+
+    }
+
+    public void logout(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if (session != null) {
+            session.invalidate();
+        }
+
+        SecurityContextHolder.clearContext();
+    }
+}
